New subproject: BoredomFS
Allows mounting a Windows machine under FUSE, somewhat like sshfs. Advantages
are that I know it doesn't waste too much RAM or otherwise do anything silly,
and it can run EXEs on the Windows box as if they have a shebang saying to run
on the remote machine.

Also the mandatory bunch of assorted updates and fixes to Arlib.